+++ src/Objects/Collection.php
@@ -3,7 +3,6 @@ namespace andrefelipe\Orchestrate\Objects;
 
 use andrefelipe\Orchestrate\Query\KeyRangeBuilder;
 use andrefelipe\Orchestrate\Query\PatchBuilder;
-use JmesPath\Env as JmesPath;
 
 /**
  * 
@@ -27,11 +26,6 @@ class Collection extends AbstractList
         return $this->getChildClass()
             ->newInstance($this->getCollection(true), $key, $ref)
             ->setClient($this->getClient(true));
-    }
-
-    public function jmesPath($expression)
-    {
-        return JmesPath::search($expression, $this->getResults());
     }
 
     /**